+++ src/index.ts
@@ -9,5 +9,6 @@ import ScoreDisplay from "src/CreditAPI/score/ScoreDisplay/ScoreDisplay";
 
 import useInsights from "./CreditAPI/insights/useInsights";
 import AccountSummary from "./CreditAPI/insights/AccountSummary/AccountSummary";
+import CreditBalance from "./CreditAPI/insights/CreditBalance/CreditBalance";
 
-export { LevelCreditProvider, useScore, ScoreChart, ScoreDisplay, useInsights, AccountSummary };
+export { LevelCreditProvider, useScore, ScoreChart, ScoreDisplay, useInsights, AccountSummary, CreditBalance };